implemented plugin authentication persistence
