+++ Assets/UI/Prefabs/Elements/Plugin Auth Panel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fc8e325a2066b5046bbe966f38a27884
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
