+++ Assets/UI/Scripts/Elements/PluginAuthEntry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f9b330ea5b6935f4ab095bbf57d09970
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
